fix(vite): don't use async entry
